refactor: move validate_google_credentials to utils.py
Separates helper function from __init__.py per review feedback.
Import path unchanged: from agno.context.google import validate_google_credentials
